requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
importlib-metadata==4.11.2
Markdown==3.3.6
oauthlib==3.2.2
Pillow==9.0.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.7.0
